requirements.txt
python-decouple
django
channels
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
pillow
psycopg2-binary
aiofiles
chainner_ext
einops
filelock
fsspec
html5tagger
httptools
Jinja2
MarkupSafe
mpmath
multidict
networkx
numpy
opencv-python
packaging
pillow
pillow-avif-plugin
pip
safetensors
sanic
sanic-routing
setuptools
spandrel
sympy
torch
torchvision
tracerite
triton
typing_extensions
ujson
uvloop
websockets
nvidia-ml-py
spandrel_extra_arches
psutil
